fixed asmall error in dev-setup
`dev-setup.md` was linking to an incorrect location for the ANet Configuration documentation. It now reflects the actual location of that file within the repository.

Note: It may be wise to switch to a relative path for these documentation files, in case they get moved. It seems like that's what happened here: Originally the documentation files were in the root directory for the project, but when they were moved to the `doc/` directory, the link broke.