NCI-Agency/anet#196 Do not allow a user to submit a report if not assigned a position
Users are only allowed to submit a report when they are assigned a
position.

If a user does not have a report they can save a draft version of the
rport. The user receeives a warning that they should contact the
administrator to be assigned a position. The user should then be able to
fully submit the report.

Changes to be committed:
	modified:   src/models/Person.js
	modified:   src/pages/reports/Form.js
	modified:   src/pages/reports/Show.js