NCI-Agency/anet#155: Fix charts when the grouping property is empty
This makes sure that if the property of the reports on which we group in
order to do the counting is empty, we reset the property to an object
representing the empty value of the property. This makes sure that we
also display a bar in the carts for 'No advisor organisation', 'No
PoAM', 'No location allocated'.